+++ R/utils_download_sianedata.R
@@ -110,5 +110,5 @@ esp_hlp_dwnload_sianedata <- function(
   if (verbose) {
     message("File loaded")
   }
-  return(err_onload)
+  err_onload
 }